Unify git-viack into viack for great pattern transformation justice
git doesn't translate Perl patterns into Vim patterns, but now git-viack
will.

To switch from the previous git-viack or a git viack alias, just
unlink/unalias it and link to viack instead using git-viack (or
git-vigrep).  The git command when viack is invoked as git-vi... is
always "grep", regardless of what "..." is.